// rpd-client-ts/src/pages/teacher-interface/model/pathes.ts
export enum TemplatePagesPath {
  COVER_PAGE = "coverPage",
  APPROVAL_PAGE = "approvalPage",
  AIMS_PAGE = "aimsPage",
  DISCIPLINE_PLACE = "disciplinePlace",
  DISCIPLINE_PLANNED_RESULTS = "disciplinePlannedResults",
  DISCIPLINE_SCOPE = "disciplineScope",
  DISCIPLINE_CONTENT = "disciplineContent",
  DISCIPLINE_SUPPORT = "disciplineSupport",
  DISCIPLINE_EVALUATIONS_FUNDS = "disciplineEvaluationsFunds",
  RESOURCE_SUPPORT = "resourceSupport",
  TEST_PDF = "testPdf",
}